Changed packages to fall under adaptors [#886]
