+++ comixed-adaptors/src/main/java/org/comixedproject/adaptors/ComicInfoEntryAdaptor.java
@@ -29,8 +29,8 @@ import javax.xml.stream.XMLStreamWriter;
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.lang.time.DateFormatUtils;
-import org.comixedproject.loaders.AbstractEntryLoader;
-import org.comixedproject.loaders.EntryLoaderException;
+import org.comixedproject.adaptors.loaders.AbstractEntryLoader;
+import org.comixedproject.adaptors.loaders.EntryLoaderException;
 import org.comixedproject.model.comicbooks.Comic;
 import org.comixedproject.model.comicbooks.Credit;
 import org.springframework.stereotype.Component;
