+++ comixed-adaptors/src/main/java/org/comixedproject/adaptors/archive/AbstractArchiveAdaptor.java
@@ -29,10 +29,10 @@ import org.apache.commons.compress.utils.IOUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.codehaus.plexus.util.FileUtils;
 import org.comixedproject.adaptors.ComicInfoEntryAdaptor;
-import org.comixedproject.handlers.ComicFileHandler;
-import org.comixedproject.handlers.ComicFileHandlerException;
-import org.comixedproject.loaders.EntryLoader;
-import org.comixedproject.loaders.EntryLoaderException;
+import org.comixedproject.adaptors.handlers.ComicFileHandler;
+import org.comixedproject.adaptors.handlers.ComicFileHandlerException;
+import org.comixedproject.adaptors.loaders.EntryLoader;
+import org.comixedproject.adaptors.loaders.EntryLoaderException;
 import org.comixedproject.model.comicbooks.Comic;
 import org.comixedproject.utils.ComicFileUtils;
 import org.comixedproject.utils.FileTypeIdentifier;
