+++ comixed-adaptors/src/main/java/org/comixedproject/adaptors/archive/SevenZipArchiveAdaptor.java
@@ -29,7 +29,7 @@ import org.apache.commons.compress.archivers.sevenz.SevenZArchiveEntry;
 import org.apache.commons.compress.archivers.sevenz.SevenZFile;
 import org.apache.commons.compress.archivers.sevenz.SevenZMethod;
 import org.apache.commons.compress.archivers.sevenz.SevenZOutputFile;
-import org.comixedproject.loaders.EntryLoaderException;
+import org.comixedproject.adaptors.loaders.EntryLoaderException;
 import org.comixedproject.model.archives.ArchiveType;
 import org.comixedproject.model.comicbooks.Comic;
 import org.springframework.stereotype.Component;
