+++ comixed-adaptors/src/main/java/org/comixedproject/adaptors/archive/ZipArchiveAdaptor.java
@@ -32,7 +32,7 @@ import org.apache.commons.compress.archivers.ArchiveStreamFactory;
 import org.apache.commons.compress.archivers.zip.ZipArchiveEntry;
 import org.apache.commons.compress.archivers.zip.ZipArchiveOutputStream;
 import org.apache.commons.compress.archivers.zip.ZipFile;
-import org.comixedproject.loaders.EntryLoaderException;
+import org.comixedproject.adaptors.loaders.EntryLoaderException;
 import org.comixedproject.model.archives.ArchiveType;
 import org.comixedproject.model.comicbooks.Comic;
 import org.springframework.stereotype.Component;
