+++ comixed-adaptors/src/main/java/org/comixedproject/adaptors/handlers/ComicFileHandler.java
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses>
  */
 
-package org.comixedproject.handlers;
+package org.comixedproject.adaptors.handlers;
 
 import java.io.*;
 import java.util.*;
