+++ comixed-adaptors/src/main/java/org/comixedproject/adaptors/loaders/FilenameEntryLoader.java
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses>
  */
 
-package org.comixedproject.loaders;
+package org.comixedproject.adaptors.loaders;
 
 import java.io.File;
 import java.util.ArrayList;
