+++ comixed-adaptors/src/test/java/org/comixedproject/adaptors/handlers/ComicFileHandlerTest.java
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses>
  */
 
-package org.comixedproject.handlers;
+package org.comixedproject.adaptors.handlers;
 
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertSame;
