+++ comixed-adaptors/src/test/java/org/comixedproject/adaptors/loaders/ImageEntryLoaderTest.java
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses>
  */
 
-package org.comixedproject.loaders;
+package org.comixedproject.adaptors.loaders;
 
 import static org.junit.Assert.*;
