+++ comixed-batch/src/main/java/org/comixedproject/batch/comicbooks/processors/ComicInsertProcessor.java
@@ -19,7 +19,7 @@
 package org.comixedproject.batch.comicbooks.processors;
 
 import lombok.extern.log4j.Log4j2;
-import org.comixedproject.handlers.ComicFileHandler;
+import org.comixedproject.adaptors.handlers.ComicFileHandler;
 import org.comixedproject.model.comicbooks.Comic;
 import org.comixedproject.model.comicfiles.ComicFileDescriptor;
 import org.comixedproject.service.comicbooks.ComicService;
