+++ comixed-batch/src/test/java/org/comixedproject/batch/comicbooks/processors/UpdateMetadataProcessorTest.java
@@ -23,8 +23,8 @@ import static junit.framework.TestCase.assertSame;
 
 import org.comixedproject.adaptors.archive.ArchiveAdaptor;
 import org.comixedproject.adaptors.archive.ArchiveAdaptorException;
+import org.comixedproject.adaptors.handlers.ComicFileHandler;
 import org.comixedproject.batch.BatchException;
-import org.comixedproject.handlers.ComicFileHandler;
 import org.comixedproject.model.archives.ArchiveType;
 import org.comixedproject.model.comicbooks.Comic;
 import org.junit.Test;
