+++ comixed-opds/src/main/java/org/comixedproject/opds/OPDSController.java
@@ -27,8 +27,8 @@ import lombok.extern.log4j.Log4j2;
 import marvin.image.MarvinImage;
 import marvinplugins.MarvinPluginCollection;
 import org.comixedproject.adaptors.archive.ArchiveAdaptorException;
+import org.comixedproject.adaptors.handlers.ComicFileHandler;
 import org.comixedproject.auditlog.AuditableEndpoint;
-import org.comixedproject.handlers.ComicFileHandler;
 import org.comixedproject.model.comicbooks.Comic;
 import org.comixedproject.model.comicbooks.Page;
 import org.comixedproject.repositories.comicbooks.ComicRepository;
