+++ comixed-rest/src/main/java/org/comixedproject/controller/comicbooks/ComicController.java
@@ -25,9 +25,9 @@ import java.util.List;
 import lombok.extern.log4j.Log4j2;
 import org.comixedproject.adaptors.archive.ArchiveAdaptor;
 import org.comixedproject.adaptors.archive.ArchiveAdaptorException;
+import org.comixedproject.adaptors.handlers.ComicFileHandler;
+import org.comixedproject.adaptors.handlers.ComicFileHandlerException;
 import org.comixedproject.auditlog.AuditableEndpoint;
-import org.comixedproject.handlers.ComicFileHandler;
-import org.comixedproject.handlers.ComicFileHandlerException;
 import org.comixedproject.model.comicbooks.Comic;
 import org.comixedproject.model.comicbooks.Page;
 import org.comixedproject.model.net.comicbooks.MarkComicsDeletedRequest;
