+++ comixed-rest/src/test/java/org/comixedproject/controller/comicbooks/ComicControllerTest.java
@@ -25,8 +25,8 @@ import java.io.InputStream;
 import java.util.List;
 import org.comixedproject.adaptors.archive.ArchiveAdaptor;
 import org.comixedproject.adaptors.archive.ArchiveAdaptorException;
-import org.comixedproject.handlers.ComicFileHandler;
-import org.comixedproject.handlers.ComicFileHandlerException;
+import org.comixedproject.adaptors.handlers.ComicFileHandler;
+import org.comixedproject.adaptors.handlers.ComicFileHandlerException;
 import org.comixedproject.model.archives.ArchiveType;
 import org.comixedproject.model.comicbooks.Comic;
 import org.comixedproject.model.comicbooks.Page;
