+++ comixed-services/src/test/java/org/comixedproject/service/comicfiles/ComicFileServiceTest.java
@@ -26,8 +26,8 @@ import java.util.ArrayList;
 import java.util.List;
 import org.comixedproject.adaptors.archive.ArchiveAdaptor;
 import org.comixedproject.adaptors.archive.ArchiveAdaptorException;
-import org.comixedproject.handlers.ComicFileHandler;
-import org.comixedproject.handlers.ComicFileHandlerException;
+import org.comixedproject.adaptors.handlers.ComicFileHandler;
+import org.comixedproject.adaptors.handlers.ComicFileHandlerException;
 import org.comixedproject.model.comicbooks.Comic;
 import org.comixedproject.model.comicfiles.ComicFile;
 import org.comixedproject.model.comicfiles.ComicFileDescriptor;
